+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### 1.4.4
 - Bump to jwt 1.5
 - Enforce hmac algorithm for jwt
+- Add AuthProvider.min_complexity, .required_chars
 
 #### 1.4.3
 - Add AuthProvider.min_length